refactor(api): rename authentication-service-registry
into oidc-authentication-service-registry which it really is.
