+++ api/config/setup-oidc-authentication-service-registry.js
@@ -1,7 +1,7 @@
-import { oidcAuthenticationServiceRegistry } from '../lib/domain/services/authentication/authentication-service-registry.js';
 import { CnavOidcAuthenticationService } from '../lib/domain/services/authentication/cnav-oidc-authentication-service.js';
 import { FwbOidcAuthenticationService } from '../lib/domain/services/authentication/fwb-oidc-authentication-service.js';
 import { GoogleOidcAuthenticationService } from '../lib/domain/services/authentication/google-oidc-authentication-service.js';
+import { oidcAuthenticationServiceRegistry } from '../lib/domain/services/authentication/oidc-authentication-service-registry.js';
 import { PaysdelaloireOidcAuthenticationService } from '../lib/domain/services/authentication/paysdelaloire-oidc-authentication-service.js';
 import { PoleEmploiOidcAuthenticationService } from '../lib/domain/services/authentication/pole-emploi-oidc-authentication-service.js';
 import { logger } from '../src/shared/infrastructure/utils/logger.js';
