+++ api/tests/unit/domain/services/authentication/oidc-authentication-service-registry_test.js
@@ -1,5 +1,5 @@
 import { InvalidIdentityProviderError } from '../../../../../lib/domain/errors.js';
-import { oidcAuthenticationServiceRegistry } from '../../../../../lib/domain/services/authentication/authentication-service-registry.js';
+import { oidcAuthenticationServiceRegistry } from '../../../../../lib/domain/services/authentication/oidc-authentication-service-registry.js';
 import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { catchErrSync, expect, sinon } from '../../../../test-helper.js';
