+++ api/tests/unit/domain/usecases/get-all-identity-providers_test.js
@@ -2,17 +2,17 @@ import { getAllIdentityProviders } from '../../../../lib/domain/usecases/get-all
 import { expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | get-all-identity-providers', function () {
-  it('returns oidc providers from authenticationServiceRegistry', function () {
+  it('returns oidc providers from oidcAuthenticationServiceRegistry', function () {
     // given
     const oneOidcProviderService = {};
     const anotherOidcProviderService = {};
-    const authenticationServiceRegistryStub = {
+    const oidcAuthenticationServiceRegistryStub = {
       getAllOidcProviderServices: sinon.stub().returns([oneOidcProviderService, anotherOidcProviderService]),
     };
 
     // when
     const identityProviders = getAllIdentityProviders({
-      authenticationServiceRegistry: authenticationServiceRegistryStub,
+      oidcAuthenticationServiceRegistry: oidcAuthenticationServiceRegistryStub,
     });
 
     // then
